Fixes for progress and level bars
The line width for the outline of the progress bar can now be set with
the options. It was previously to 'thick', now the default is 0.2 mm.
The line thickness is now also used for the symbols in the level bars.
The total width of the progress bar now also considers this thickness.
This was previously 'accounted for' by scaling the x axis. It will now
be properly subtracted.