Documentation for new sidebar line
The user-level commands for the sidebar line and the corresponding skill
and language templates have been addded to the manual.
The options for the hint width and the indentation have also been added.
The progressbar linewdith option is documented and described.
The examples have been recompiled to check for inconsistencies.

This commit should properly complete #7 (after necessary further testing).