Release Candidate v1.1rc1 (#8)
* Added line template for sidebar

Added To allow for better customisation of the side bar (#7) a line template.
Removed an internal macro and replaced it with several user-level
macros. Options to control the new lengths have been added.

* Fixes for progress and level bars

The line width for the outline of the progress bar can now be set with
the options. It was previously to 'thick', now the default is 0.2 mm.
The line thickness is now also used for the symbols in the level bars.
The total width of the progress bar now also considers this thickness.
This was previously 'accounted for' by scaling the x axis. It will now
be properly subtracted.

* Documentation for new sidebar line

The user-level commands for the sidebar line and the corresponding skill
and language templates have been added to the manual.
The options for the hint width and the indentation have also been added.
The progressbar linewidth option is documented and described.
The examples have been recompiled to check for inconsistencies.

* Update version and date, recompile and adapted examples, push #7 to RC